requirements.txt
# Qwen2.5-VL Training Requirements

# Core ML dependencies
# NOTE: Install PyTorch separately with CUDA support before running pip install -r requirements.txt
# See README.md for installation instructions
transformers>=4.46.0
peft>=0.13.0
accelerate>=0.34.0
bitsandbytes>=0.41.0

# Unsloth for efficient training
unsloth>=2024.1

# TRL for supervised fine-tuning
trl>=0.22.0

# Image processing
Pillow>=10.0.0

# Dataset handling
datasets>=4.3.0

# Hugging Face Hub
huggingface_hub>=0.34.0

# Evaluation dependencies
scikit-learn>=1.0.0
tqdm>=4.60.0
